LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-010-expected.txt


PASS Should use the last fallback position initially
FAIL Should use the third fallback position with enough space right assert_equals: Anchored element should be at the right of anchor expected 650 but got 450
FAIL Should use the second fallback position with enough space above assert_equals: Anchored element should be at the top of anchor expected 135 but got 335
FAIL Should use the first fallback position with enough space right and above assert_equals: Anchored element should be at the right of anchor expected 199 but got -1
